requirements.txt
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
openapi-codec==1.3.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.7
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
